Bug 1674353 - Do not assume ArrayBuffer's length is a valid wasm heap length. r=rhunt
In bug 1673818 an intermediary getter was introduced for the buffer
length, in anticipation of larger buffers and a distinction between
32-bit and 64-bit wasm memories.  This getter asserts that the value
it gets is below the limit for the appropriate memory type, eg,
ByteLength32 asserts that the value it gets is below MaxMemory32Bytes,
which is currently 2^31-64K.  However, ArrayBuffers up to 2^31-1 can
be constructed and passed to asm.js, so asm.js must not use this
getter until after it has validated that the buffer length is below
MaxMemory32Bytes.

Thus we use the unchecked byteLength() accessor on the buffer, and
introduce an additional guard in IsValidAsmJSHeapLength that the
length is also a valid wasm length.  This will have no impact on
asm.js in SpiderMonkey as the largest valid asm.js length is already
below the wasm maximum. (The largest valid asm.js length is the
largest value below 2^31-1 whose low 24 bits are all zero.)

Differential Revision: https://phabricator.services.mozilla.com/D95563

+++ js/src/wasm/AsmJS.h
@@ -106,7 +106,7 @@ extern JSString* AsmJSModuleToString(JSContext* cx, JS::Handle<JSFunction*> fun,
 
 // asm.js heap:
 
-extern bool IsValidAsmJSHeapLength(uint32_t length);
+extern bool IsValidAsmJSHeapLength(size_t length);
 
 }  // namespace js
